Add missing component files (#713)
* Add missing export files
