+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Added missing component files instead of `index`](https://github.com/multiversx/mx-sdk-dapp/pull/713)
 
 ## [[v2.10.13](https://github.com/multiversx/mx-sdk-dapp/pull/708)] - 2023-03-27
 - [Optimized sign-message reducers](https://github.com/multiversx/mx-sdk-dapp/pull/707)
