+++ src/UI/CopyButton/copyButton.styles.scss
