--- src/UI/NotificationModal/notificationModalStyles.scss
@@ -1,3 +1,0 @@
-.notification-modal {
-  color: inherit;
-}
