--- src/UI/walletConnect/WalletConnectLoginButton/walletConnectLoginButtonStyles.scss
@@ -1,3 +1,0 @@
-.wallect-connect-login {
-  color: inherit;
-}
